==> Logistic Algo Execution/Codes/4_1_model_fitting.ipynb <==
import os 
import numpy as np
import pandas as pd
import datetime as dt
from sklearn.linear_model.logistic import LogisticRegression

Fucntions of this notebook:
* Read indicator data ("ticker_indicator.csv" in directory ".../Data/indicator", generated by 2_tech_indicator_calculation)
* Split sample into training set and test set
* Fit two logistic regression models, one with all indicators and one with indicators selected by stepwise process (generated by 3_stepwise_selection)
* Output predicted direction and probability for test set as one .csv file (with name "ticker_signal.csv" and stored in directory ".../Output")

### Directory Setting

dir_working = "/Users/user/Desktop/E4733 AT/Project/Coding Environment/Codes"
os.chdir(dir_working)
dir_data = "../Data/indicator"
dir_output = "../Output"
#!# set ticker list 
ticker_list = ['aapl', 'amzn', 'nvda', 'amd', 'msft', 'fb', 'nflx', \
               'goog', 'intc', 'pypl']

### 1. Indicator Sets Selected by Stepwise Process

optimal_set = [['MA_CAT_15','MA_CAT_30','VMA_Cat_3','VMA_Cat_15', \
                'VMA_Cat_20','MACD_CAT_3_5','MACD_CAT_5_15','MACD_CAT_10_15',\
                'MACD_CAT_15_30','WilliamR','RSI_20','PSY_15','PSY_30','ADO',\
                'K_D_CAT_5_3','K_D_CAT_B_15_80_20'],\
               ['MA_CAT_5','MA_CAT_15','MA_CAT_20','MA_CAT_30','VMA_Cat_3',\
                'VMA_Cat_5','VMA_Cat_10','VMA_Cat_20','MACD_CAT_3_5',\
                'MACD_CAT_5_10','MACD_CAT_15_30','WilliamR','RSI_10','PSY_10',\
                'ADO','K_D_CAT_5_3','K_D_CAT_B_5_80_20'],\
               ['MA_CAT_20','VMA_Cat_3','VMA_Cat_10','VMA_Cat_15',\
                'VMA_Cat_20','MACD_CAT_3_5','MACD_CAT_5_15','WilliamR',\
                'RSI_10','RSI_20','RSI_30','PSY_10','ADO','K_D_CAT_5_3',\
                'K_D_CAT_B_5_80_20','K_D_CAT_B_15_80_20'],\
               ['MA_CAT_3','MA_CAT_5','VMA_Cat_15','MACD_CAT_3_5',\
                'MACD_CAT_5_30','MACD_CAT_10_15','WilliamR','PSY_10','PSY_15',\
                'PSY_30','ADO','K_D_CAT_15_3','K_D_CAT_5_3',\
                'K_D_CAT_B_15_80_20','CCI_5'],\
               ['MA_CAT_3','MA_CAT_10','VMA_Cat_20','MACD_CAT_5_15',\
                'WilliamR','PSY_10','ADO','K_D_CAT_15_3'],\
               ['MA_CAT_3','VMA_Cat_3','VMA_Cat_5','VMA_Cat_15',\
                'MACD_CAT_3_5','MACD_CAT_5_10','MACD_CAT_10_20','WilliamR',\
                'RSI_20','PSY_10','ADO','K_D_CAT_15_3','K_D_CAT_B_15_80_20',\
                'CCI_5'],\
               ['MA_CAT_3','MA_CAT_5','MA_CAT_15','MA_CAT_20','VMA_Cat_15',\
                'MACD_CAT_3_5','MACD_CAT_10_15','WilliamR','RSI_10','PSY_10',\
                'ADO'],\
               ['VMA_Cat_3','VMA_Cat_5','VMA_Cat_30','MACD_CAT_3_5',\
                'MACD_CAT_5_10','MACD_CAT_5_30','MACD_CAT_10_15','WilliamR',\
                'RSI_10','PSY_10','ADO','K_D_CAT_5_3','K_D_CAT_B_5_80_20',\
                'K_D_CAT_B_15_80_20','CCI_5'],\
               ['MA_CAT_3','MA_CAT_30','VMA_Cat_10','MACD_CAT_3_5',\
                'MACD_CAT_5_15','MACD_CAT_5_30','WilliamR','RSI_10','PSY_10',\
                'ADO','K_D_CAT_5_3','K_D_CAT_B_5_80_20','K_D_CAT_B_15_80_20'],\
               ['MA_CAT_15','VMA_Cat_5','MACD_CAT_3_5','MACD_CAT_5_15',\
                'MACD_CAT_5_30','MACD_CAT_10_15','WilliamR','RSI_10','PSY_10',\
                'ADO','K_D_CAT_5_3','K_D_CAT_B_15_80_20']]

### 2. Model Function 

def model_fit_optimal(ticker = 'aapl',indicator_set = optimal_set[0]):
    ## read & clean data
    data = pd.read_csv(dir_data + '/' + ticker + '_indicator.csv')
    data.time = pd.to_datetime(data.time)
    data = data.set_index('time', drop = True)
    data.direction = data.direction.shift(-1)
    data['return'] = data['return'].shift(-1)
    data = data.dropna()
    ## training & test sets 
    data['date'] = data.index.date
    # training set 
    data_train = data.where(data.date < dt.date(2018,12,17))
    data_train = data_train.dropna()
    data_train = data_train.drop('date', axis = 1)
    # test set 
    data_test = data.where(data.date >= dt.date(2018,12,17))
    data_test = data_test.dropna()
    data_test = data_test.drop('date', axis = 1)
    
    ## Logistic Full Model
    # values 
    X_train = data_train.loc[:,list(data_train.columns)[11:]].values
    Y_train = data_train.loc[:,'direction'].values
    X_test = data_test.loc[:,list(data_train.columns)[11:]].values
    Y_test = data_test.loc[:,'direction'].values
    # model
    logistic_model = LogisticRegression(solver = 'lbfgs', max_iter = 1000) 
    logistic_model.fit(X_train, Y_train) 
    acc_in_logistic_full = np.mean(logistic_model.predict(X_train) == Y_train)
    Y_pred_logistic_full = logistic_model.predict(X_test)
    acc_out_logistic_full = np.mean(Y_pred_logistic_full == Y_test)
    logistic_proba_full = logistic_model.predict_proba(X_test)
    
    ## Logistic Stepwise Optimized 
    # values 
    X_train = data_train.loc[:,indicator_set].values
    Y_train = data_train.loc[:,'direction'].values
    X_test = data_test.loc[:,indicator_set].values
    Y_test = data_test.loc[:,'direction'].values
    # model
    logistic_model = LogisticRegression(solver = 'lbfgs', max_iter = 1000) 
    logistic_model.fit(X_train, Y_train) 
    acc_in_logistic = np.mean(logistic_model.predict(X_train) == Y_train)
    Y_pred_logistic = logistic_model.predict(X_test)
    acc_out_logistic = np.mean(Y_pred_logistic == Y_test)
    logistic_proba = logistic_model.predict_proba(X_test)
    
    ## Output & Return 
    data_test['signal_logistic_full'] = Y_pred_logistic_full
    data_test['logistic_prob_full'] = np.max(logistic_proba_full, axis = 1)
    data_test['signal_logistic'] = Y_pred_logistic
    data_test['logistic_prob'] = np.max(logistic_proba, axis = 1)
    data_output = data_test.drop(list(data_train.columns)[11:], axis = 1)
    data_output.to_csv(dir_output + '/' + ticker + '_signal.csv')
    return [acc_in_logistic_full,acc_out_logistic_full,\
            acc_in_logistic,acc_out_logistic]

### 3. Execution 

for i in range(len(ticker_list)):
    summary_list = model_fit_optimal(ticker_list[i])
    print('\n'+ticker_list[i])
    print('Full Model:')
    print('in-sample: ', summary_list[0])
    print('out-of-sample: ', summary_list[1])
    print('Stepwise Optimized:')
    print('in-sample: ', summary_list[2])
    print('out-of-sample: ', summary_list[3], '\n')
